+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/beneficiary/BeneficiaryDetail.kt
@@ -62,14 +62,19 @@ import kotlinx.coroutines.launch
 @Composable
 fun BeneficiaryDetailScreen(
     onBackClick: () -> Unit,
-    onModifyClick: () -> Unit,
-    viewModel: BeneficiaryDetailViewModel = hiltViewModel()
+    onModifyClick: (String) -> Unit,
+    viewModel: BeneficiaryDetailViewModel = hiltViewModel(),
+    beneficiaryId: String,
 ) {
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
     val snackbarHostState = remember { SnackbarHostState() }
     val scope = rememberCoroutineScope()
 
     var showDeleteDialog by remember { mutableStateOf(false) }
+
+    LaunchedEffect(Unit) {
+        viewModel.loadBeneficiary(beneficiaryId)
+    }
 
     LaunchedEffect(uiState.deleteSuccess, uiState.deleteError) {
         if (uiState.deleteSuccess) {
@@ -92,14 +97,14 @@ fun BeneficiaryDetailScreen(
         snackbarHostState = snackbarHostState,
         showDeleteDialog = showDeleteDialog,
         onBackClick = onBackClick,
-        onModifyClick = {
-            uiState.beneficiary?.id?.let { onModifyClick() }
-        },
+        onModifyClick = onModifyClick,
         onDeleteClick = viewModel::onDeleteClicked,
         onShowDialog = { showDeleteDialog = true },
-        onDismissDialog = { showDeleteDialog = false }
+        onDismissDialog = { showDeleteDialog = false },
+        beneficiaryId = beneficiaryId,
     )
 }
+
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun BeneficiaryDetailContent(
@@ -107,10 +112,11 @@ fun BeneficiaryDetailContent(
     snackbarHostState: SnackbarHostState,
     showDeleteDialog: Boolean,
     onBackClick: () -> Unit,
-    onModifyClick: () -> Unit,
+    onModifyClick: (String) -> Unit,
     onDeleteClick: () -> Unit,
     onShowDialog: () -> Unit,
     onDismissDialog: () -> Unit,
+    beneficiaryId: String,
     sessionViewModel: SessionViewModel = hiltViewModel(),
 ) {
     val role by sessionViewModel.role.collectAsState()
@@ -166,13 +172,15 @@ fun BeneficiaryDetailContent(
             }
             uiState.beneficiary != null -> {
                 val beneficiary = uiState.beneficiary
+                val imageUrl = beneficiary.image.takeIf { !it.isNullOrBlank() }
+
                 Column(
                     modifier = Modifier
                         .padding(paddingValues)
                         .padding(horizontal = 24.dp)
                         .fillMaxSize()
                 ) {
-                   Column(
+                    Column(
                         modifier = Modifier
                             .weight(1f)
                             .verticalScroll(rememberScrollState()),
@@ -190,7 +198,7 @@ fun BeneficiaryDetailContent(
                             ) {
                                 Image(
                                     painter = rememberAsyncImagePainter(
-                                        model = beneficiary.image,
+                                        model = imageUrl,
                                         placeholder = painterResource(id = R.drawable.ic_beneficiary_icon),
                                         error = painterResource(id = R.drawable.img_arca_logo)
                                     ),
@@ -225,7 +233,7 @@ fun BeneficiaryDetailContent(
                             }
                         }
 
-                       Spacer(modifier = Modifier.height(16.dp))
+                        Spacer(modifier = Modifier.height(16.dp))
 
                         Row(
                             modifier = Modifier.fillMaxWidth(),
@@ -238,7 +246,7 @@ fun BeneficiaryDetailContent(
                                 TextValue(label = "Número de teléfono", value = beneficiary.emergencyNumber.orEmpty())
                             }
                         }
-                       Spacer(modifier = Modifier.height(16.dp))
+                        Spacer(modifier = Modifier.height(16.dp))
 
                         Row(
                             modifier = Modifier.fillMaxWidth(),
@@ -279,10 +287,10 @@ fun BeneficiaryDetailContent(
 
                         Spacer(modifier = Modifier.width(16.dp))
 
-                        /*ModifyButton(
+                        ModifyButton(
                             modifier = Modifier.size(width = 112.dp, height = 40.dp),
-                            onClick = onModifyClick
-                        )*/
+                            onClick = { onModifyClick(beneficiaryId) }
+                        )
                     }
                 }
             }
@@ -331,7 +339,6 @@ fun DetailTextRow(label: String, value: String?) {
 }
 
 
-
 @Preview(name = "Activo", showBackground = true, backgroundColor = 0xFF1C1B1F)
 @Composable
 fun BeneficiaryDetailPreviewActive() {
@@ -359,7 +366,8 @@ fun BeneficiaryDetailPreviewActive() {
             onModifyClick = {},
             onDeleteClick = {},
             onShowDialog = {},
-            onDismissDialog = {}
+            onDismissDialog = {},
+            beneficiaryId = ""
         )
     }
 }
@@ -380,8 +388,8 @@ fun BeneficiaryDetailPreviewInactive() {
                     emergencyRelation = "Padre",
                     details = "Detalles del beneficiario",
                     entryDate = "01/01/2023",
-                    image = "",
-                    disabilities = listOf("Si"),
+                    image = "https://ajemplo.com",
+                    disabilities =  listOf("Visual"),
                     status = 0
                 )
             ),
@@ -391,7 +399,8 @@ fun BeneficiaryDetailPreviewInactive() {
             onModifyClick = {},
             onDeleteClick = {},
             onShowDialog = {},
-            onDismissDialog = {}
+            onDismissDialog = {},
+            beneficiaryId = ""
         )
     }
 }