+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/user/register/UserRegisterScreen.kt
@@ -1,6 +1,7 @@
 package com.app.arcabyolimpo.presentation.screens.user.register
 
 import android.annotation.SuppressLint
+import android.net.Uri
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
@@ -18,6 +19,7 @@ import com.app.arcabyolimpo.presentation.ui.components.atoms.buttons.SaveButton
 import com.app.arcabyolimpo.presentation.ui.components.atoms.buttons.CancelButton
 import com.app.arcabyolimpo.presentation.ui.components.atoms.icons.ExitIcon
 import com.app.arcabyolimpo.presentation.ui.components.atoms.inputs.ModalInput
+import com.app.arcabyolimpo.presentation.ui.components.atoms.inputs.ImageUploadInput
 import androidx.compose.foundation.clickable
 import com.app.arcabyolimpo.presentation.ui.components.atoms.icons.CalendarIcon
 import androidx.compose.material3.SnackbarHost
@@ -37,6 +39,9 @@ fun UserRegisterScreen(
     onSuccess: () -> Unit
 ) {
     val uiState by viewModel.uiState.collectAsState()
+    // COLLECTED STATE FROM VIEWMODEL (for image preview)
+    val selectedImageUri by viewModel.selectedImageUri.collectAsState()
+
     val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
     val isActiveBoolean = uiState.isActive == 1
 
@@ -50,9 +55,15 @@ fun UserRegisterScreen(
                 snackbarHostState.showSnackbar(message)
             }
         } else if (uiState.error != null) {
-
             uiState.error?.let { errorMessage ->
-                snackbarHostState.showSnackbar(errorMessage)
+                // Usamos el mensaje de error directamente
+                scope.launch {
+                    snackbarHostState.showSnackbar(
+                        message = errorMessage,
+                        withDismissAction = true,
+                        duration = SnackbarDuration.Long
+                    )
+                }
             }
         }
     }
@@ -191,6 +202,16 @@ fun UserRegisterScreen(
                         .verticalScroll(rememberScrollState()),
                     verticalArrangement = Arrangement.spacedBy(16.dp)
                 ) {
+                    // Image Upload
+                    ImageUploadInput(
+                        label = "Foto de Perfil",
+                        value = selectedImageUri,
+                        onValueChange = { uri ->
+                            viewModel.setSelectedImageUri(uri)
+                        },
+                        height = 200.dp
+                    )
+
                     // Role Selection
                     Column(modifier = Modifier.fillMaxWidth()) {
                         Text(