+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 ## [0.2.0] - 02/12/2025
+### Added
+- Consult disability details
+
 ### Changed
 - Attendance screen now always displays exit time as 14:30.
+  
+## [0.2.0] - 01/12/2025
+### Added
+- Cache for product list and detail
+- Modify beneficiary
+- Add functionality to add, view, and modify images in Beneficiary, Workshop, Supply, Product and User
+
+### Changed
+ - Modify topbar for scaffold
+
+### Fixed
+- Debug modify beneficiary
 
 ## [0.2.0] - 30/11/2025
 ### Added
@@ -17,6 +32,7 @@
 ### Added
 - Added offline for product batch list and detail
 - Added a Capacitations view for Disabilities List
+- Added offline for supply batches
 
 ## [0.2.0] - 29/11/2025